drafting the ocsp tests, see #6
this is an attempt to test the OCSP stapling in gmid.  Support for it
was already added to gg, the test now only loads gmid with a custom
certificate and use the (new) -O flag of gg.

It currently fails, probably because I'm not generating the ocsp
response correctly, but it's in the way for other works on the regress
suite so it gets its own draft/* branch for the next raining day.